requirements.txt
accelerate==0.20.3
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blis==0.7.11
bs4==0.0.1
boto3==1.28.56
botocore==1.31.56
Brotli==1.1.0
catalogue==2.0.10
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
crispy-bootstrap4==2022.1
cymem==2.0.8
datasets==2.14.6
dill==0.3.7
dj-database-url==2.1.0
Django==3.2.21
django-crispy-forms==2.0
django-storages==1.14
dnspython==2.4.2
evaluate==0.4.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gunicorn==21.2.0
huggingface==0.0.1
huggingface-hub==0.17.3
idna==3.4
imageio==2.31.6
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
langcodes==3.3.0
lazy_loader==0.3
LinkChecker==10.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack-python==0.5.6
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
packaging==23.1
pandas==2.1.2
pathy==0.10.3
Pillow==9.5.0
preshed==3.0.9
psutil==5.9.6
psycopg2-binary==2.9.8
pyarrow==14.0.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
redis==2.10.6
regex==2023.10.3
requests==2.31.0
responses==0.18.0
s3transfer==0.7.0
safetensors==0.4.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
sympy==1.12
thinc==8.2.1
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.14.1
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
weasel==0.3.3
whitenoise==6.5.0
xxhash==3.4.1
yarl==1.9.2
